Tugas 1

+++ Tugas-1/Soal-2.py
@@ -1,5 +1,5 @@
 Pi = 22 / 7
-r = int(input("Masukkan jari-jari lingkaran (cm): "))
+r = float(input("Masukkan jari-jari lingkaran (cm): "))
 Luas = float(Pi * r ** 2)
 
 text = "Luas lingkaran dengan jari-jari {} cm adalah {:.2f} cm\u00b2".format (r, Luas)
